backend/requirements.txt
Flask==2.1.0
python-dotenv==0.19.2
supabase==1.0.3
requests==2.26.0
anthropic==0.3.0
# Add Speechify API library when available